Move MockBuildTemporaryEnvironment to build_helper.py
