+++ rebasehelper/build_tools/mock_tool.py
@@ -25,24 +25,9 @@ import os
 from rebasehelper.utils import ProcessHelper
 from rebasehelper.logger import logger
 from rebasehelper.utils import PathHelper
-from rebasehelper.build_helper import BuildTemporaryEnvironment
 from rebasehelper.build_helper import BuildToolBase
 from rebasehelper.build_helper import BinaryPackageBuildError
-
-
-class MockTemporaryEnvironment(BuildTemporaryEnvironment):
-    """
-    Class representing temporary environment for MockBuildTool.
-    """
-
-    def _create_directory_structure(self):
-        # create directory structure
-        for dir_name in ['SOURCES', 'SPECS', 'RESULTS']:
-            self._env[self.TEMPDIR + '_' + dir_name] = os.path.join(
-                self._env[self.TEMPDIR], dir_name)
-            logger.debug("Creating '%s'",
-                         self._env[self.TEMPDIR + '_' + dir_name])
-            os.makedirs(self._env[self.TEMPDIR + '_' + dir_name])
+from rebasehelper.build_helper import MockTemporaryEnvironment
 
 
 class MockBuildTool(BuildToolBase):  # pylint: disable=abstract-method
